feat(texture): add texture to batch renderer (#11)
* feat: add Component and GameObject

* feat: BatchRenderer, not finished yet

* refactor: add comment

* feat: add BatchRenderer and renamed classes

* feat: add BatchScene, no error but still nothing visible

* fix: camera, batch still broken

* feat: batch rendering now working :)

* fix: changed group name in build.gradle.kts

* refactor: move Component to components dir

* feat: element component and batch renderer (#8)

* feat: add Component and GameObject

* feat: BatchRenderer, not finished yet

* refactor: add comment

* feat: add BatchRenderer and renamed classes

* feat: add BatchScene, no error but still nothing visible

* fix: camera, batch still broken

* feat: batch rendering now working :)

* fix: changed group name in build.gradle.kts

* refactor: move Component to components dir

* feat: add assetpool

* feat: replaced shader and texture creation by asset provider func

* feat: add asset provider to manage resources (#10)

* feat: add Component and GameObject

* feat: BatchRenderer, not finished yet

* refactor: add comment

* feat: add BatchRenderer and renamed classes

* feat: add BatchScene, no error but still nothing visible

* fix: camera, batch still broken

* feat: batch rendering now working :)

* fix: changed group name in build.gradle.kts

* refactor: move Component to components dir

* feat: add assetpool

* feat: replaced shader and texture creation by asset provider func

* fix: changed applicationStartTime type from double to long

* fix: centered camera and adjusted obj position

* fix: return after found scene by keypress instead of continuing to iterate

* feat: changed levelScene to have a square gameobject

* refactor: remove empty line

* refactor: removed unused methods in camera

* feat: placed new object in LevelScene

* refactor: rename

* feat: extracted TextureProvider and ShaderProvider

* feat: add texture and id to vertexBuffer

* fix: binding to slot

* feat: created LocalBatchTextureProvider to hold to textures referenced by ids to current batch

* feat: added width and height to Camera

+++ src/main/java/io/nova/core/utils/Time.java
@@ -1,7 +1,7 @@
 package io.nova.core.utils;
 
 public class Time {
-    public static double applicationStartTime = System.nanoTime();
+    public static long applicationStartTime = System.nanoTime();
 
     public static double getElapsedTimeSinceApplicationStartInSeconds() {
         return (System.nanoTime() - applicationStartTime) * 1E-9;
